LPS-89210 adding, updating Layout creates new LayoutVersion, model listeners checking null layout when deleting non-persisted layout drafts

+++ modules/apps/subscription/subscription-service/src/main/java/com/liferay/subscription/internal/model/listener/SubscriptionLayoutModelListener.java
@@ -32,6 +32,10 @@ public class SubscriptionLayoutModelListener extends BaseModelListener<Layout> {
 
 	@Override
 	public void onAfterRemove(Layout layout) {
+		if (layout == null) {
+			return;
+		}
+
 		try {
 			_subscriptionLocalService.deleteSubscriptions(
 				layout.getCompanyId(), Layout.class.getName(),
